+++ ARKBreedingStats/ARKBreedingStats.csproj
@@ -106,15 +106,15 @@
     <Compile Include="mods\StatBaseValuesEdit.Designer.cs">
       <DependentUpon>StatBaseValuesEdit.cs</DependentUpon>
     </Compile>
-    <Compile Include="ocr\Common\Pattern.cs" />
-    <Compile Include="ocr\Common\TextData.cs" />
-    <Compile Include="ocr\Common\DirectBitmap.cs" />
-    <Compile Include="ocr\Common\ImageUtils.cs" />
-    <Compile Include="ocr\Common\PatternOcr.cs" />
-    <Compile Include="ocr\Common\RecognitionPatterns.cs" />
-    <Compile Include="ocr\Common\CoordsData.cs" />
-    <Compile Include="ocr\Common\Coords.cs" />
-    <Compile Include="ocr\Common\RecognizedCharData.cs" />
+    <Compile Include="ocr\PatternMatching\Pattern.cs" />
+    <Compile Include="ocr\PatternMatching\TextData.cs" />
+    <Compile Include="ocr\PatternMatching\DirectBitmap.cs" />
+    <Compile Include="ocr\PatternMatching\ImageUtils.cs" />
+    <Compile Include="ocr\PatternMatching\PatternOcr.cs" />
+    <Compile Include="ocr\PatternMatching\RecognitionPatterns.cs" />
+    <Compile Include="ocr\PatternMatching\CoordsData.cs" />
+    <Compile Include="ocr\PatternMatching\Coords.cs" />
+    <Compile Include="ocr\PatternMatching\RecognizedCharData.cs" />
     <Compile Include="oldLibraryFormat\FormatConverter.cs" />
     <Compile Include="importExported\ExportedCreatureList.cs">
       <SubType>Form</SubType>
