+++ ARKBreedingStats/RecognitionTrainingForm.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Drawing;
 using System.Windows.Forms;
-using ARKBreedingStats.ocr.Common;
+using ARKBreedingStats.ocr.PatternMatching;
 
 namespace ARKBreedingStats
 {
