+++ ARKBreedingStats/ocr/PatternMatching/CoordsData.cs
