+++ ARKBreedingStats/ocr/PatternMatching/DirectBitmap.cs
